split cmake targets, include paths, isolate boost dependency to mpl implementation
