+++ CMakeLists.txt
@@ -2,38 +2,25 @@ cmake_minimum_required(VERSION 3.14)
 
 set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
 
-project(compile_time_graph_shortest_path VERSION 1.0.0 LANGUAGES CXX)
+project(compile_time_graph_bfs VERSION 1.0.0 LANGUAGES CXX)
 
-# Set C++ standard
-set(CMAKE_CXX_STANDARD 20)
-set(CMAKE_CXX_STANDARD_REQUIRED ON)
+find_package(Boost CONFIG REQUIRED)
 
-# Find Boost
-find_package(Boost REQUIRED)
-
-# Add include directory for the header-only library
-add_library(compile_time_graph_shortest_path INTERFACE)
-target_include_directories(compile_time_graph_shortest_path INTERFACE
-    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>
+add_library(consteval_graph_bfs INTERFACE)
+target_include_directories(consteval_graph_bfs INTERFACE
+    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>/consteval_graph_bfs/include
     $<INSTALL_INTERFACE:include>
 )
-target_link_libraries(compile_time_graph_shortest_path INTERFACE Boost::boost)
+
+add_library(mpl_graph_bfs INTERFACE)
+target_include_directories(mpl_graph_bfs INTERFACE
+    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>/mpl_graph_bfs/include
+    $<INSTALL_INTERFACE:include>
+)
+target_link_libraries(mpl_graph_bfs INTERFACE Boost::headers)
 
 # Testing
 option(BUILD_TESTING "Build tests" ON)
 if(BUILD_TESTING)
-    enable_testing()
-    
-    # Fetch Google Test
-    include(FetchContent)
-    FetchContent_Declare(
-        googletest
-        GIT_REPOSITORY https://github.com/google/googletest.git
-        GIT_TAG v1.14.0
-    )
-    # For Windows: Prevent overriding the parent project's compiler/linker settings
-    set(gtest_force_shared_crt ON CACHE BOOL "" FORCE)
-    FetchContent_MakeAvailable(googletest)
-    
     add_subdirectory(tests)
 endif()
